+++ firestore-bigquery-export/firestore-bigquery-change-tracker/src/bigquery/partitioning.ts
@@ -53,7 +53,8 @@ export class Partitioning {
 
   private isValidPartitionTypeDate(value) {
     /* Check if valid timestamp value from sdk */
-    if (value instanceof firebase.firestore.Timestamp) return true;
+    // if (value instanceof firebase.firestore.Timestamp) return true;
+    if (isTimestampLike(value)) return true;
 
     /* Check if valid date/timstemap, expedted result from production  */
     if (value && value.toDate && value.toDate()) return true;
@@ -88,15 +89,19 @@ export class Partitioning {
       !timePartitioningField &&
       !timePartitioningFieldType &&
       !timePartitioningFirestoreField;
-
     /* No custom config has been set, use partition value option only */
     if (hasNoCustomOptions) return true;
 
-    /* check if all options have been provided to be  */
+    /* check if all valid combinations have been provided*/
+    const hasOnlyTimestamp =
+      timePartitioningField === "timestamp" &&
+      !timePartitioningFieldType &&
+      !timePartitioningFirestoreField;
     return (
-      !!timePartitioningField &&
-      !!timePartitioningFieldType &&
-      !!timePartitioningFirestoreField
+      hasOnlyTimestamp ||
+      (!!timePartitioningField &&
+        !!timePartitioningFieldType &&
+        !!timePartitioningFirestoreField)
     );
   }
 
@@ -123,30 +128,24 @@ export class Partitioning {
   }
 
   hasValidTableReference() {
-    logs.invalidTableReference();
+    if (!this.table) {
+      logs.invalidTableReference();
+    }
     return !!this.table;
   }
 
   private async isTablePartitioned() {
-    if (!this.table) return Promise.resolve(false);
-    // No table provided, cannot evaluate
-    if (this.table.exists()) {
-      logs.cannotPartitionExistingTable(this.table);
-      return Promise.resolve(false);
-    }
-
-    /*** No table exists, return */
-    const [tableExists] = await this.table.exists();
-    if (!tableExists) return Promise.resolve(false);
-
-    /* Check if partition metadata already exists */
+    /* Return true if partition metadata already exists */
     const [metadata] = await this.table.getMetadata();
-    if (!!metadata.timePartitioning) return Promise.resolve(true);
+    if (!!metadata.timePartitioning) {
+      logs.cannotPartitionExistingTable(this.table);
+      return Promise.resolve(true);
+    }
 
     /** Find schema fields **/
     const schemaFields = await this.metaDataSchemaFields();
 
-    /** No Schema exists, return */
+    /** Return false if no schema exists */
     if (!schemaFields) return Promise.resolve(false);
 
     /* Return false if time partition field not found */
@@ -156,14 +155,12 @@ export class Partitioning {
   }
 
   async isValidPartitionForExistingTable(): Promise<boolean> {
+    /** Return false if partition type option has not been set */
+    if (!this.isPartitioningEnabled()) return Promise.resolve(false);
+
+    /* Return false if table is already partitioned */
     const isPartitioned = await this.isTablePartitioned();
     if (isPartitioned) return Promise.resolve(false);
-
-    return this.hasValidCustomPartitionConfig();
-  }
-
-  isValidPartitionForNewTable(): boolean {
-    if (!this.isPartitioningEnabled()) return false;
 
     return this.hasValidCustomPartitionConfig();
   }
@@ -189,7 +186,7 @@ export class Partitioning {
     Extracts a valid Partition field from the Document Change Event.
     Matches result based on a pre-defined Firestore field matching the event data object.
     Return an empty object if no field name or value provided. 
-    Returns empty object if not a string or timestamp
+    Returns empty object if not a string or timestamp (or result of serializing a timestamp)
     Logs warning if not a valid datatype
     Delete changes events have no data, return early as cannot partition on empty data.
   **/
@@ -210,6 +207,15 @@ export class Partitioning {
 
     if (this.isValidPartitionTypeDate(fieldValue)) {
       /* Return converted console value */
+      if (isTimestampLike(fieldValue)) {
+        const convertedTimestampFieldValue = convertToTimestamp(fieldValue);
+        return {
+          [fieldName]: this.convertDateValue(
+            convertedTimestampFieldValue.toDate()
+          ),
+        };
+      }
+
       if (fieldValue.toDate) {
         return { [fieldName]: this.convertDateValue(fieldValue.toDate()) };
       }
@@ -237,63 +243,66 @@ export class Partitioning {
   }
 
   async addPartitioningToSchema(fields = []): Promise<void> {
-    /** check if class has valid table reference */
-    if (!this.hasValidTableReference()) return Promise.resolve();
+    /** Return if partition type option has not been set */
+    if (!this.isPartitioningEnabled()) return;
 
-    /** return if table is already partitioned **/
-    if (await this.isTablePartitioned()) return Promise.resolve();
+    /** Return if class has invalid table reference */
+    if (!this.hasValidTableReference()) return;
 
-    /** return if an invalid partition type has been requested**/
-    if (!this.hasValidTimePartitionType()) return Promise.resolve();
+    /** Return if table is already partitioned **/
+    if (await this.isTablePartitioned()) return;
+
+    /** Return if partition config is invalid */
+    if (!this.hasValidCustomPartitionConfig()) return;
+
+    /** Return if an invalid partition type has been requested */
+    if (!this.hasValidTimePartitionType()) return;
+
+    /** Return if an invalid partition option has been requested */
+    if (!this.hasValidTimePartitionOption()) return;
 
     /** Return if invalid partitioning and field type combination */
-    if (this.hasHourAndDatePartitionConfig()) return Promise.resolve();
+    if (this.hasHourAndDatePartitionConfig()) return;
 
-    /** return if an invalid partition type has been requested**/
-    if (!this.hasValidCustomPartitionConfig()) return Promise.resolve();
+    /** Return if partition field has not been provided */
+    if (!this.config.timePartitioningField) return;
 
-    /** return if an invalid partition type has been requested**/
-    if (!this.hasValidCustomPartitionConfig()) return Promise.resolve();
+    /** Return if field already exists on schema */
+    if (this.customFieldExists(fields)) return;
 
-    /** update fields with new schema option ** */
-    if (!this.hasValidTimePartitionOption()) return Promise.resolve();
-
-    /* Check if partition field has been provided */
-    if (!this.config.timePartitioningField) return Promise.resolve();
-
-    // if (await !this.hasExistingSchema) return Promise.resolve();
-
-    // Field already exists on schema, skip
-    if (this.customFieldExists(fields)) return Promise.resolve();
-
+    /** Add new partitioning field **/
     fields.push(getNewPartitionField(this.config));
 
-    /** log successful addition of partition column */
+    /** Log successful addition of partition column */
     logs.addPartitionFieldColumn(
       this.table.id,
       this.config.timePartitioningField
     );
 
-    return Promise.resolve();
+    return;
   }
 
   async updateTableMetadata(options: bigquery.TableMetadata): Promise<void> {
-    /** return if table is already partitioned **/
-    if (await this.isTablePartitioned()) return Promise.resolve();
+    /** Return if partition type option has not been set */
+    if (!this.isPartitioningEnabled()) return;
 
-    /** return if an invalid partition type has been requested**/
-    if (!this.hasValidTimePartitionType()) return Promise.resolve();
+    /** Return if class has invalid table reference */
+    if (!this.hasValidTableReference()) return;
 
-    /** update fields with new schema option ** */
-    if (!this.hasValidTimePartitionOption()) return Promise.resolve();
+    /** Return if table is already partitioned **/
+    if (await this.isTablePartitioned()) return;
+
+    /** Return if an invalid partition type has been requested**/
+    if (!this.hasValidCustomPartitionConfig()) return;
+
+    /** Return if an invalid partition type has been requested**/
+    if (!this.hasValidTimePartitionType()) return;
+
+    /** Update fields with new schema option ** */
+    if (!this.hasValidTimePartitionOption()) return;
 
     /** Return if invalid partitioning and field type combination */
-    if (this.hasHourAndDatePartitionConfig()) return Promise.resolve();
-
-    /** return if an invalid partition type has been requested**/
-    if (!this.hasValidCustomPartitionConfig()) return Promise.resolve();
-
-    // if (await !this.hasExistingSchema) return Promise.resolve();
+    if (this.hasHourAndDatePartitionConfig()) return;
 
     if (this.config.timePartitioning) {
       options.timePartitioning = { type: this.config.timePartitioning };
@@ -309,3 +318,27 @@ export class Partitioning {
     }
   }
 }
+
+type TimestampLike = {
+  _seconds: number;
+  _nanoseconds: number;
+};
+
+const isTimestampLike = (value: any): value is TimestampLike => {
+  if (value instanceof firebase.firestore.Timestamp) return true;
+  return (
+    typeof value === "object" &&
+    value !== null &&
+    "_seconds" in value &&
+    typeof value["_seconds"] === "number" &&
+    "_nanoseconds" in value &&
+    typeof value["_nanoseconds"] === "number"
+  );
+};
+
+const convertToTimestamp = (
+  value: TimestampLike
+): firebase.firestore.Timestamp => {
+  if (value instanceof firebase.firestore.Timestamp) return value;
+  return new firebase.firestore.Timestamp(value._seconds, value._nanoseconds);
+};