+++ firestore-bigquery-export/CHANGELOG.md
@@ -1,3 +1,40 @@
+## Version 0.1.44
+
+fix - apply the task queue configurations in `syncBigQuery` and set max attempts to 5
+fix - MAX_DISPATCHES_PER_SECOND is now set to take effect based on user's configuration
+
+## Version 0.1.43
+
+fix - correctly partition when only "timestamp" is selected for partition options
+
+## Version 0.1.42
+
+fix - correctly extract timestamps from firestore fields to partition columns
+
+## Version 0.1.41
+
+fix - rollback backfill feature
+
+## Version 0.1.40
+
+fix - correct default value for use collection group query param
+
+## Version 0.1.39
+
+fix - rollback timestamp serialization
+
+## Version 0.1.38
+
+fix - backfill value mismatch
+
+## Version 0.1.37
+
+fix - serialize timestamps to date string
+
+## Version 0.1.36
+
+build - updated depenencies
+
 ## Version 0.1.35
 
 fixed - add missing locations back in
@@ -65,6 +102,8 @@ fixed - update snapshot script
 feature - add oldData to the record
 
 fixed - updating table metadata too often
+
+feature - add lifecycle event to export existing documents to Bigquery
 
 ## Version 0.1.26